adição da classe controller e regras de negócio

+++ src/main/java/com/example/SistemaBlblioteca/service/BookService.java
@@ -3,10 +3,14 @@ package com.example.SistemaBlblioteca.service;
 import com.example.SistemaBlblioteca.dto.livroDTO.BookRequestDTO;
 import com.example.SistemaBlblioteca.dto.livroDTO.BookResponseDTO;
 import com.example.SistemaBlblioteca.entity.Book;
+import com.example.SistemaBlblioteca.entity.Category;
 import com.example.SistemaBlblioteca.exceptions.BookAlreadyExistException;
 import com.example.SistemaBlblioteca.exceptions.BookNotFoundException;
+import com.example.SistemaBlblioteca.exceptions.BookUnavailableException;
+import com.example.SistemaBlblioteca.exceptions.CategoryNotFoundException;
 import com.example.SistemaBlblioteca.mapper.BookMapper;
 import com.example.SistemaBlblioteca.repository.BookRepository;
+import com.example.SistemaBlblioteca.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +19,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class BookService {
+    private final CategoryRepository categoryRepository;
     private final BookRepository bookRepository;
     private final BookMapper bookMapper;
 
@@ -22,9 +27,22 @@ public class BookService {
         if(bookRepository.existsByTitle(bookDTO.title())) {
             throw new BookAlreadyExistException("Book with name: " +bookDTO.title()+ " already exists");
         }
-        Book book = bookMapper.toEntity(bookDTO);
-        bookRepository.save(book);
-        return bookMapper.toDTO(book);
+
+        Category category = categoryRepository.findById(bookDTO.categoryId())
+                .orElseThrow(() -> new CategoryNotFoundException("Category with id: " +bookDTO.categoryId()+ " was not found"));
+
+        Book book = new  Book();
+        book.setTitle(bookDTO.title());
+        book.setCategory(category);
+        book.setAvailable(true);
+        Book savedBook = bookRepository.save(book);
+
+        return new BookResponseDTO(
+                savedBook.getId(),
+                savedBook.getTitle(),
+                savedBook.getCategory().getName(),
+                savedBook.isAvailable()
+        );
     }
 
     public BookResponseDTO findByTitle(String title) {
@@ -43,10 +61,17 @@ public class BookService {
     }
 
     public BookResponseDTO update(Long id, BookRequestDTO bookDTO) {
+        Category category = categoryRepository.findById(bookDTO.categoryId())
+                .orElseThrow(() -> new CategoryNotFoundException("Category with id: " +bookDTO.categoryId()+ " was not found"));
+
         Book book = getBook(id);
+        if (!book.isAvailable()) {
+            throw new BookUnavailableException("Unavailable books cannot be updated");
+        }
 
         book.setTitle(bookDTO.title());
-        book.setCategory(bookDTO.category_id());
+        book.setCategory(category);
+        book.setAvailable(true);
 
         bookRepository.save(book);
         return bookMapper.toDTO(book);
@@ -54,6 +79,10 @@ public class BookService {
 
     public void delete(Long id) {
         Book book = getBook(id);
+        if (!book.isAvailable()) {
+            throw new BookUnavailableException("Unavailable books cannot be deleted");
+        }
+
         bookRepository.delete(book);
     }
